CDP-1561 - move press packages data to separate mock file
